feat(skills): box `pay skills ls` like search
Extract the 80-column bordered-box renderer (frame + wrap + indent) into
a shared `skills::boxed` module and use it for both commands:

- `search` keeps its endpoint sections (output unchanged), now via the
  shared helpers instead of local copies.
- `ls` renders each provider as a framed section — title + endpoint
  count, then the dimmed fqn (+ pinned marker) — matching the search look.
