+++ rust/crates/cli/src/commands/skills/mod.rs
@@ -1,3 +1,4 @@
+mod boxed;
 pub mod endpoints;
 pub mod install;
 pub mod list;
